requirement.txt
fastapi==0.61.1
uvicorn==0.12.1
pymongo==3.11.0
minio==6.0.0
pillow==8.0.0
pybase64==1.0.2